Lift branch (#9)
* Lift Subsystem

Lift Subsystem was defined in this branch.

* Changes to lift subsystem

Just making it look a little nicer and I added a TODO

* Changed Little Thing

Drive Base Stuff Boi, like the master

* New DriveWithLift Command

I made a command to drive the lift using a joystick axis.

* Test At Pull Requests

Test 1

* New ShiftLift Command

I created a new ShiftLift Command and assigned it on button Y of the XBox controller.

* Addition of Position Lift Functions

I created functions to use the encoders in order to position the lift to three different values. I still need to figure out the values so... it's not done COMPLETELY.

* Adding On To Collector Things

Andie created some more things for the collector

* Changes to LiftWithEncoders Method and Command

Now the command takes in parameters!!!!!

* Finished the Lift!

The lift subsystem is now complete but untested (TODO test this).

* Changes To Lift PID Things

I setup the lift PID stuff wrong at first, so I fixed it (hopefully, I haven't really done this type of thing before).