Diagnostics (#14)
* Making the Master Nice

We basically cleaned up everything and made it compile

* Testing Out Things

We have to put this together with the master, but this is the most updated version as of now.

* Cleaning and Niceness

This is the most recent version, I basically "messed with" a few things and fixed a logic problem for the collector.

* Making FMS Code

It was a little bit of a logic problem at first, but I think I did it.

* Version 1.0 Diagnostics

This is the first edition of the Diagnostics code!
Made by Connor Nicholls

* Version 1.1

DO NOT MERGE YET
This is an attempt to see if combining OITest and OI will work. If not we will revert this.

* Version 1.2

Everything is OK now! Let me test this to make sure though. Note: Diagnostics is now on button 7 on the logitech so the drivers don't start clearing sticky faults in a match

* Version 1.2.1

This just clarifies the SmartDashboard output.

* bugfixing

fixing some random bugs with live window and cims

* more bugfixing (hopefully the last one)

fixed some more top cim issues